refactoring: comments7_1 (Resource)
