+++ application/src/main/java/org/thingsboard/server/controller/TbResourceController.java
@@ -40,7 +40,7 @@ import org.thingsboard.server.common.data.page.PageData;
 import org.thingsboard.server.common.data.page.PageLink;
 import org.thingsboard.server.common.data.security.Authority;
 import org.thingsboard.server.queue.util.TbCoreComponent;
-import org.thingsboard.server.service.entitiy.resource.TbResourceNotifyService;
+import org.thingsboard.server.service.resource.TbResourceService;
 import org.thingsboard.server.service.security.permission.Operation;
 import org.thingsboard.server.service.security.permission.Resource;
 
@@ -71,7 +71,7 @@ import static org.thingsboard.server.controller.ControllerConstants.UUID_WIKI_LI
 @RequiredArgsConstructor
 public class TbResourceController extends BaseController {
 
-    private  final TbResourceNotifyService tbResourceService;
+    private  final TbResourceService tbResourceService;
 
     public static final String RESOURCE_ID = "resourceId";
