+++ application/src/main/java/org/thingsboard/server/service/resource/TbResourceService.java
@@ -24,12 +24,13 @@ import org.thingsboard.server.common.data.id.TenantId;
 import org.thingsboard.server.common.data.lwm2m.LwM2mObject;
 import org.thingsboard.server.common.data.page.PageData;
 import org.thingsboard.server.common.data.page.PageLink;
+import org.thingsboard.server.service.entitiy.SimpleTbEntityService;
 
 import java.util.List;
 
-public interface TbResourceService {
+public interface TbResourceService extends SimpleTbEntityService<TbResource> {
 
-    TbResource saveResource(TbResource resource) throws ThingsboardException;
+    TbResource saveResourceInternal(TbResource resource) throws ThingsboardException;
 
     TbResource getResource(TenantId tenantId, ResourceType resourceType, String resourceKey);
 
@@ -50,8 +51,6 @@ public interface TbResourceService {
                                           String sortProperty,
                                           String sortOrder,
                                           PageLink pageLink);
-
-    void deleteResource(TenantId tenantId, TbResourceId resourceId);
 
     void deleteResourcesByTenantId(TenantId tenantId);
